Gradients work shader switches between solid colour and texture lookup defines for stock shader states SWF sRGB correct Math has addef sRGB convert function from D3DX Scene:: Map/Unmap function moved down to resource level (now both textures and databuffers can be mapped, updated etc.) Copy/Update function added to update GPU resources 64bit supported (in theory) for Swf files, though memory allocates would fail on 32 bit

+++ src/libs/gui/swfparser/SwfButton.cpp
@@ -16,7 +16,7 @@ namespace Swf
 {
 	SwfButtonObject* SwfButtonObject::Read( SwfStream& _stream, int _length, int _version ) {
 		
-		long marker = _stream.marker();
+		uint64_t marker = _stream.marker();
 		
 		uint16_t id = _stream.readUInt16();
 
@@ -63,7 +63,8 @@ namespace Swf
 
 		// version one only had one action press and release
 		if( _version == 1 ) {
-			_length -= (_stream.marker() - marker);
+			// TODO 64 bit files
+			_length -= (int)(_stream.marker() - marker);
 		
 			SwfActionByteCode* actionScript = CORE_NEW SwfActionByteCode();
 			actionScript->byteCode = CORE_NEW_ARRAY uint8_t[_length];
